Implement grouping
Now the parser returns a vector of selector chains instead of one. All
the functions that deal with selector chains have been adjusted to
account for this change. When selecting, the controller iterates over
the multiple selector chains and creates a selection for each, and then
the outer vectors are merged.