Info.plist - properly export UTI for the HSS file type and correct icon references and add URL schemes.
This means: HSS icons display "Hierarchial Style Sheet" as the file type in Finder, are displayed using the prototype's icon, and display as a text file in Quick Look. Also, the application advertises that it supports http, https and file URL schemes. This doesn't really do anything yet but will be ready for when it does.