Prepend AXR_ to almost all of our CMake option() commands.
This avoids potential conflicts with built-in variables and gives a nice
overview of the available options when using the GUI as they will be
listed alphabetically.

BUILD_UNIVERSAL_BINS is close enough in name and function class to the
built-in BUILD_SHARED_LIBS, so it can stay as it is.