Introduce AXRString. This will be the library's main string class.
All std::string instances have been replaced with AXRString, except in
the SDL application due to boost::program_options. Deal with that later.

Currently it's simply a typedef for QString but eventually we will
have it encapsulate QString and duplicate its API in order to keep
external dependencies out of our header files.

Path manipulations through substring monkeying have been replaced with
QUrl and QFileInfo where appropriate, which should be a lot safer.

This commit also "breaks" Unicode support, because the original solution
where HSSTextBlock blindly interpreted its contents as UTF-8 is
incorrect. What if a document is in UTF-16? Shift-JIS? ISO-8859-1?
This needs to be handled accordingly at an earlier stage according to
either the encoding of the opened document itself or the encoding
attribute of the XML prolog, or both. We'll have to find a way to deal
with documents that don't specify an encoding (such as directly opening
HSS files) - ISO-8859-1 or UTF-8 come to mind. We'll have to deal with
the UTF-8 BOM as well.