Use HSSUnit across the entire codebase and fix serious rendering quirks.
Explanation: in 3ed7b11, the HSSUnit typedef was changed from long
double to double, as the former is not consistent across platforms (it
may be 64 or 80 bits dependent on the compiler and other factors).
Unfortunately, due to the way that a lot of HSSUnit values are handled
in the codebase (void* and such), this led to an inconsistency where
some values were doubles and some were long doubles, and caused
expressions such as 50% + 20 to evaluate to 20, and presumably had
other unknown side effects as well. This commit uses HSSUnit across the
entire codebase and solves the problem. It also seems to have alleviated
a problem where XML text nodes "were not loading properly" - in
actuality it was rendering issues due to these type issues.