Rewrite all file loading to use QUrl for network transparency and being able to better handle the differences between relative and absolute paths.
Transforming relative to absolute paths is done in two places: XMLParser
(as any HSS files loaded from there are relative to the XML file itself)
and in HSSParser::readNextStatement when reading #import instructions
(as any HSS files loaded from there are relative to the current HSS file).

However, why is HSSRequest loading XML files? This situation may not be
handled properly for relative paths…

AXRWrapper::loadFileByPath, loadXMLFile, and loadHSSFile, may take an
absolute or relative path. These are user-facing methods that may be
called by client applications. If a local, relative path is given it
will be interpreted relative to the application's current working
directory.

AXRWrapper::getFile accepts absolute paths only. If a relative path is
given to it, an error will be thrown. This is an internal method that
should not be called externally, and thus any paths passed to it should
be converted to absolute paths.

Hardcoding "file://" + path and all sorts of other hacks is very error
prone, and caused the Windows prototype not to load files at all, which
this issue also fixes.

Some files are included for testing that relative paths are loaded
properly, however due to #154 this is somewhat difficult to test
properly so there may be regressions introduced with this commit.

The AXRTestThread and AXRTestProducer classes may be broken now, but
they are such a mess that they need to be completely redone anyways.

Files now load properly on Windows!

Closes #147.