Fix .rc file for SDL app and add one for main prototype app so that the apps will have proper icons on Windows.
I've also re-saved all the .ico files. Visual Studio 2010's resource
compiler was giving the RC2176 error when attempting to compile the
prototype and example executables. For some reason the old files were
corrupt somehow, with Visual Studio failing to display all but the
256x256 icon, showing sizes such as 48x0 in menus, as well as other
error messages.

I've saved the .ico files with Icon Composer (just as before) but this
time followed a very specific procedure:

* File -> New ICO
* Drag 1024x1024 PNG from Finder into 48x48 slot
* Drag 1024x1024 PNG from Finder into 32x32 slot
* Drag 1024x1024 PNG from Finder into 24x24 slot
* Drag 1024x1024 PNG from Finder into 16x16 slot
* Drag 1024x1024 PNG from Finder into 256x256 slot
* Save file

I believe I originally dragged the icon from Finder to the 256x256 slot,
and then from the 256x256 slot into each of the others from largest to
smallest size. Apparently the order and/or location from which icons
are dragged do make a difference in Icon Composer although inspecting a
binary diff did appear to reveal a few minor differences while the order
of the icons themselves remained the same.