Fix rendering of borders.
In the process of updating to our new visitor pattern for drawing, some old code got moved and the newer one deleted. This restores the call to the shape, which will be in charge of actually rendering the border. Since we are now working in a global coordinate system, the offset needs to be passed in. Also, since we are no longer working in constrained surfaces, the borderBleeding is not needed anymore.