Fix issue that caused #new elements inside an object definition to be inserted twice
Since we are adding the cloned rule to the parent rule, when returning to the
recursiveMatchRulesToDisplayObjects() method it will be matched again. Now the
question arises where to put these rules. Until now they were applied before
all the others, which is the same than we do now. This is probably not the
definitive behavior, though.